requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
better-profanity==0.5.0
bidict==0.22.1
blinker==1.6.3
blis==0.7.11
boto3==1.28.71
botocore==1.31.71
CacheControl==0.12.14
cachetools==4.2.4
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==1.6.0
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
Command==0.1.0
confection==0.1.3
confluent-kafka==1.9.2
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
databases==0.8.0
dnspython==2.3.0
docopt==0.6.2
eventlet==0.33.3
exceptiongroup==1.1.3
fastapi==0.104.1
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gevent==23.9.1
google-auth==1.35.0
google-auth-oauthlib==0.5.3
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.19.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jax==0.4.13
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.12
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.3.4
mattermostwrapper==2.2
ml-dtypes==0.2.0
monotonic==1.5
msgpack==1.0.7
multidict==5.2.0
murmurhash==1.0.10
mysql-connector-python==8.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pathy==0.10.3
Pillow==10.1.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pymssql==2.2.9
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.1.1
pyreadline3==3.4.1
pyrsistent==0.20.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2021.3
pywin32==306
PyYAML==6.0.1
questionary==1.10.0
